create new dto and entities
